LayoutTests/imported/w3c/web-platform-tests/2dcontext/drawing-images-to-the-canvas/2d.drawImage.animated.poster-expected.txt

2d.drawImage.animated.poster
drawImage() of an APNG draws the poster frame
Actual output:
Expected output:


FAIL drawImage() of an APNG draws the poster frame assert_approx_equals: Red channel of the pixel at (50, 25) expected 0 +/- 2 but got 255
